- SHE macro protection - Fix missing CMAC wolfCrypt typecasts - Add utils module for network functions
