+++ test/wh_test_she.h
@@ -2,5 +2,6 @@
 #define WH_TEST_SHE_H_
 
 int whTest_She(void);
+int whTest_SheClientConfig(whClientConfig* config);
 
 #endif /* WH_TEST_SHE_H_ */
